Task4/Iterate 1
I instantiate the ElectricHitEffect in my SimpleProjectile script the moment it hits something damageable at the position the porjectile was when it hit. I also added a Selfdestroyer script to the ElectricHitEffect with a countdown which either destroy the effect after the effect duration and if the gameobject atttached to this sript has no particle system it destroys it after the countown.