Task5/Iterate 2(v1)
Added the PickUpText by subscribing to the OnCollected Event. Everytime the palyer collects a drop the OnPickUpLoot(Drop drop) gets called now. Which displayes the PickupText predefined by the designer in the scriptableobject drop. PickUptext. The method fades the text out by lowere its text color aalpha to 0 and shrinking it by altering its font size over time. if a new Drop gets collected meanwhile the procces gets stopped immediately. The text will be restted and the whole process begins for the new object.
Also added 2 editable speed values for the fading and shrinking speed in the Canvas_PlayerHUD.