Task6/Prototype 1 Special Attack
To pickup the Special Blast. I adjusted the sripütableobject drop to identifiy as an ability (bool) and hav ae projectile spell desription member. Now on the OnCollected event I check if my Drops is an Ability and fire the OnabilityCollected Event. Which enables the UI for the 2. Ability slot with the values of the 2. Spells and also enables the Spellcastincontroller to be able to use right click for the second ability. The methods to actually use the second ability are all similiar to the light attack to add another collectable ability to the drops you have to: create a drop with the ability bool markes and add a projectile description to it. The Projectile description is also a scriptable object you just have to create (just duplicate the Basic prefab) and adjust the values accordingly. Last you just have to add the drop to the pool of dropable object from the salamander and decide on a probability.

The Lootprobability was also altered in the process now it correctly looks through all drops in the pool and iterates over them until only one gets returned and droped on  the screen.